requirements.txt
wandb
regex
hydra-core
ptwt
termcolor
ftfy
torch
torchvision
torchaudio
pandas
torchmetrics[image]
omegaconf
einops
timm
huggingface_hub
datasets==2.18.0
treescope
accelerate
evaluate
loguru
lightning
ruff
black
isort